fixed the comments
