+++ apps/web/.eslintrc.json
@@ -1,17 +1,9 @@
 {
 	"extends": ["next/core-web-vitals", "next/typescript"],
 	"rules": {
-	  "no-unused-vars": ["warn", { 
-		"argsIgnorePattern": "^_",
-		"varsIgnorePattern": "^_",
-		"destructuredArrayIgnorePattern": "^_"
-	  }],
-	  "@typescript-eslint/no-unused-vars": ["warn", { 
-		"argsIgnorePattern": "^_",
-		"varsIgnorePattern": "^_",
-		"destructuredArrayIgnorePattern": "^_"
-	  }],
-	  "@typescript-eslint/no-empty-object-type": "warn",
-	  "react/no-unescaped-entities": "warn"
+		"no-unused-vars": "warn",
+		"@typescript-eslint/no-unused-vars": "warn",
+		"@typescript-eslint/no-empty-object-type": "warn",
+		"react/no-unescaped-entities": "warn"
 	}
-  }
+}
