+++ apps/web/components/shared/kyc/kyc-4/document-preview.tsx
