+++ apps/web/components/shared/kyc/kyc-4/extracted-info-display.tsx
