+++ apps/web/components/shared/kyc/kyc-4/types.ts
@@ -1,0 +1,7 @@
+export type DocumentType = 'utility' | 'bank' | 'government' | ''
+
+export interface ExtractedData {
+	text: string
+	date: string | null
+	address: string | null
+}
