+++ packages/lib/package.json
@@ -10,5 +10,4 @@
 		"tesseract.js": "^6.0.0"
 	},
 	"devDependencies": {}
-	
 }
