+++ packages/lib/src/doc-utils/extraction.ts
