+++ packages/lib/src/index.ts
@@ -1,3 +1,3 @@
-export * from './doc-utils/upload-handler'
 export * from './doc-utils/validation'
 export * from './doc-utils/extraction'
+export * from './doc-utils/file-processing'
